Add rpn.py (early WIP)
I'm now considering re-implementing rpn in Python.

It turns out Python's support for arbitrary size integers is better
than Perl's.  In Perl, I have to check whether a stack item is a scalar
(number), a BigInt, or a BigFloat.  In Python, built-in integers
expand on demand.

The Perl version has a `-man` option, using `Pod::Usage`.
I'm not yet sure how to do the equivalent in Python.